Move CrashReporting/Breadcrumb from support-base to concept-base.
