+++ components/service/firefox-accounts/src/test/java/mozilla/components/service/fxa/AccountStorageTest.kt
@@ -6,7 +6,7 @@ package mozilla.components.service.fxa
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.lib.dataprotect.SecureAbove22Preferences
-import mozilla.components.support.base.crash.CrashReporting
+import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
