+++ components/support/migration/src/main/java/mozilla/components/support/migration/FennecLoginsMPImporter.kt
@@ -5,7 +5,7 @@
 package mozilla.components.support.migration
 
 import mozilla.components.service.sync.logins.ServerPassword
-import mozilla.components.support.base.crash.CrashReporting
+import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.support.base.log.logger.Logger
 
 /**
