+++ components/support/migration/src/main/java/mozilla/components/support/migration/TelemetryIdentifiers.kt
@@ -4,7 +4,7 @@
 
 package mozilla.components.support.migration
 
-import mozilla.components.support.base.crash.CrashReporting
+import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.support.base.log.logger.Logger
 import org.json.JSONException
 import org.json.JSONObject
